+++ src/components/ApplicationViews.js
@@ -1,15 +1,15 @@
 import React, { Component } from "react";
 import { Route } from "react-router-dom";
-import Quests from "./quest/Quests";
-import CharacterNewForm from "./characters/CharacterNewForm";
+import CharacterForm from "./characters/CharacterForm";
+import QuestViews from "./quest/QuestViews";
 
 export default class ApplicationViews extends Component {
   render() {
     return (
       <>
         {/* <Route exact path="/" render={props => <Quests />} /> */}
-        <Route exact path="/quests" render={props => <Quests {...props} />} />
-
+        <Route path="/quests" render={props => <QuestViews {...props} />} />
+        {/* 
         <Route
           path="/quests/:questId(\d+)"
           render={props => (
@@ -17,9 +17,11 @@ export default class ApplicationViews extends Component {
           )}
         />
 
+        <Route path="/quests/new" render={props => <Quests {...props} />} /> */}
+
         <Route
           path="/character/new"
-          render={props => <CharacterNewForm {...props} />}
+          render={props => <CharacterForm {...props} />}
         />
         {/* <Route path="/guild" render={props => <Guild />} /> */}
       </>
