+++ src/components/quest/Quests.css
@@ -6,9 +6,16 @@
 }
 
 .log-sidebar {
+  display: flex;
+  flex-flow: column wrap;
   background-color: steelblue;
   margin: 0;
   padding: 0;
+  min-height: 100%;
+}
+
+.log-sidebar>* {
+  flex: 0 0 50%;
 }
 
 .quest-display {
@@ -16,11 +23,11 @@
   margin: 0;
 }
 
-.quest-list-container, .quest-details-container {
+.quest-list-container, .quest-details-container, .quest-form-container {
   height: 100%;
 }
 
-.quest-details-body, .quest-list-body {
+.quest-details-body, .quest-list-body, .quest-form-body {
   height: 100%;
 }
